+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - Resolved newsletter flag: Disable the default value to false !1779 @chriswalg
 - Resolved the user registration: Removed adding the accepted_date to getPnVersion !1779 @chriswalg
 - Disabled session cookie checkbox in login form, until there is a solution in issue 956 !1785 @chriswalg
+- Disabled session cookie checkbox in login form and, until there is a solution in issue 956
+and persistent session disabled after each request, so that LastLogin is set correctly again !1785 @chriswalg
 
 ## Refactoring
 - Flatten some workgroup function calls !1702 @ChrisOelmueller